requirements.txt
python-vlc
mutagen
PySide6
